WL#16350: Update dnspython version
Classic and XDevAPI connectors can be installed
optionally with all the 3rd-party packages
required by the dns-srv feature.

The dnspython package is part of these 3rd-party
requirements; however, the current installable
versions of this package are unsecured - they are
hit by the vulnerability `NVD CVE-2023-29483`.

This work log updates the version range such that
secure versions of dnspython are installable when
exercising the dns-srv installation option.

Change-Id: I14e895ac286b11da7a3ad550465926b363fac7f4