BUG#36731062: mysql_native_password plugin not bundled with the C-extension
Starting with version 9.0.0 the mysql_native_password client-side
plugin is required for connection to MySQL accounts that use
native password authentication. That plugin is not bundled
with the connector and hence it is impossible to make such a
connection.

With this patch, the mysql_native_password plugin is bundled
like the connector bundles other authentication plugins such
as authentication_kerberos_client and authentication_webauthn_client.

Change-Id: Ic47de6a075b8789893330a18c5839e8c87605fcd