BUG#37541353: (Contribution) Fix typing annotation of MySQLConnectionAbstract's close function
As part of the connection API, improved the type
hint of `close()` which is an alias of the instance
method `disconnect()`. Thanks, Parul Gupta, for the
contribution!

Additionally, since `close()` is the preferred term (see
[PEP 249 – Python Database API Specification v2.0]),
with this patch, we also changed roles by letting
`disconnect()` now be an alias of `close()`.

From a functional perspective, nothing changes, this
modification is oriented to provide consistency and
clarity to the interface declaration.

Change-Id: I61d407186fcc9fa5496c617f05593c4c69bdfac8